MiMi-d: Add new osc 2 detune parameter
In conjunction with this, move the sync parameter to the osc common
page, and put osc 2 detune where sync was on the osc 2 page.

The parameter numbers have been rearranged, so that osc 2 detune
takes the parameter number previously used by osc sync, and
osc sync gets a new parameter number.